carniceria_bot/requirements.txt
# Core dependencies
openai>=1.0.0
google-generativeai>=0.3.0

# Template rendering and config
Jinja2>=3.1.0
PyYAML>=6.0.0

# Web framework (for dashboard)
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# Email
python-dotenv>=1.0.0

# Google Sheets API
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1

# WhatsApp
twilio>=8.9.0

# Cache (optional)
redis>=5.0.0

# Monitoring & Error Tracking
sentry-sdk>=1.40.0

# CLI
rich>=13.7.0
click>=8.1.0

# Data Generation (for testing)
Faker>=22.0.0

# Utilities
requests>=2.31.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-benchmark>=4.0.0
coverage>=7.4.0

# Code Quality
flake8>=7.0.0
bandit>=1.7.6
black>=24.1.1

# Security
bcrypt>=4.1.2
cryptography>=42.0.0

# Multimedia
openai-whisper>=20231117
pydub>=0.25.1
soundfile>=0.12.1
